refactor: user and provider service communication

+++ provider-service/src/main/java/dev/zbib/providerservice/service/ProviderService.java
@@ -13,13 +13,11 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
+import static dev.zbib.providerservice.model.mapper.DetailsMapper.toDetailsListResponse;
 import static dev.zbib.providerservice.model.mapper.ProviderMapper.toProvider;
 import static dev.zbib.providerservice.model.mapper.ProviderMapper.toProviderListResponse;
 
@@ -30,6 +28,8 @@ public class ProviderService {
 
     private final ProviderRepository providerRepository;
     private final UserClient userClient;
+    private final UserService userService;
+
 
     public void registerProvider(
             Long id,
@@ -45,52 +45,29 @@ public class ProviderService {
 
     public void deleteProviderByUserId(Long userId) {
         Provider provider = getProviderById(userId);
-        providerRepository.delete(provider);
+        if (provider != null) {
+            providerRepository.delete(provider);
+        }
     }
 
-    public Page<ProviderListResponse> getProviders(
+    public Page<ProviderListResponse> getProviderPage(
             ServiceType serviceType,
             Boolean available,
             Double hourlyRate,
             String serviceArea,
             Pageable pageable) {
-
-        Specification<Provider> specification = ProviderSpecification.createFilter(
-                serviceType,
-                available,
-                hourlyRate,
-                serviceArea);
-
+        var specification = ProviderSpecification.createFilter(serviceType, available, hourlyRate, serviceArea);
         Page<Provider> providerPage = providerRepository.findAll(specification, pageable);
-        List<Long> userIds = providerPage.getContent()
-                .stream()
-                .map(Provider::getId)
-                .collect(Collectors.toList());
+        List<UserListResponse> users = userService.getUserDetailsForProviders(providerPage.getContent());
+        List<ProviderListResponse> providerList = toProviderListResponse(
+                providerPage.getContent(),
+                users);
 
-        List<UserListResponse> userList = userClient.getUsersByIds(userIds);
-        Map<Long, UserListResponse> userMap = userList.stream()
-                .collect(Collectors.toMap(UserListResponse::getId, user -> user));
-
-        List<ProviderListResponse> providerList = providerPage.getContent()
-                .stream()
-                .map(provider -> {
-                    UserListResponse user = userMap.get(provider.getId());
-                    return toProviderListResponse(provider, user);
-                })
-                .collect(Collectors.toList());
         return new PageImpl<>(providerList, pageable, providerPage.getTotalElements());
     }
 
-    public List<DetailsListResponse> getDetailListById(List<Long> id) {
-        List<Provider> providerList = providerRepository.findProvidersByIdIn(id);
-        return providerList.stream()
-                .map(user -> DetailsListResponse.builder()
-                        .id(user.getId())
-                        .available(user.isAvailable())
-                        .hourlyRate(user.getHourlyRate())
-                        .serviceArea(user.getServiceArea())
-                        .serviceType(user.getServiceType())
-                        .build())
-                .collect(Collectors.toList());
+    public List<DetailsListResponse> getDetailListById(List<Long> ids) {
+        List<Provider> providerList = providerRepository.findProvidersByIdIn(ids);
+        return toDetailsListResponse(providerList);
     }
 }
